# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: WOVSDH Nails Booking API
  description: |
    Complete RESTful API for online nail salon booking system with Telegram notifications, Google Calendar integration, and email confirmations.

    ## Base URLs

    - **Local Development**: `http://localhost:3000`
    - **Production**: `https://your-domain.com`

    ## Authentication

    ### Public Endpoints
    Most endpoints are public and don't require authentication. However, Supabase RLS (Row Level Security) ensures data isolation.

    ### Admin Endpoints
    Admin endpoints require `X-Admin-Secret` header with the value of `ADMIN_SECRET_KEY` environment variable.

    ### Internal Endpoints
    Internal endpoints (confirmed bookings, email sending) require `X-Internal-Secret` header with the value of `ADMIN_SECRET_KEY`.

    ## Response Format

    All responses are JSON. Errors follow a standard format:

    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message"
      }
    }
    ```

    ## Dates and Times

    All timestamps are returned in ISO 8601 format with timezone offset (e.g., `2025-06-01T08:00:00+03:00`).

    All dates in query parameters use format `YYYY-MM-DD`.

    ## Rate Limiting

    No rate limiting is currently implemented. For production use, implement rate limiting via Vercel's edge functions or external service.

    ## Webhooks

...
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/bookings:
    post:
      summary: Create a new booking
      description: |
        Create a new booking for a specific slot. The booking starts in PENDING status and requires confirmation via Telegram bot.

        The system will:
        1. Check if the slot is available
        2. Create a booking record
        3. Mark the slot as booked
        4. Send a Telegram notification to staff
        5. Generate a unique cancellation token

        A cancellation token is automatically generated and embedded in the cancellation email link.
      operationId: createBooking
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            examples:
              example1:
                value:
                  slot_id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                  studio_id: rishon
                  client_first_name: Анна
                  client_last_name: Иванова
                  client_phone: '+972501234567'
                  client_email: anna@example.com
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBookingResponse'
          examples:
            success:
              value:
                booking:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  slot_id: f47ac10b-58cc-4372-a567-0e02b2c3d479
                  studio_id: rishon
                  status: PENDING
                  start_at: '2025-06-01T08:00:00+03:00'
                  end_at: '2025-06-01T09:00:00+03:00'
                  created_at: '2025-05-18T12:34:56.000Z'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                validation_error:
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid email format
        '409':
          description: Slot is not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                unavailable:
                  value:
                    error:
                      code: SLOT_UNAVAILABLE
                      message: Slot is already booked
  /api/bookings/cancel:
    post:
      summary: Cancel a booking
      description: |
        Cancel an existing booking using the cancellation token. The token is sent to the client via email when the booking is confirmed.

        Cancellation will:
        1. Update booking status to CANCELLED
        2. Free the slot (mark as available)
        3. Delete the event from Google Calendar (if it exists)
        4. Send a cancellation email to the client
      operationId: cancelBooking
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
            examples:
              example1:
                value:
                  token: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelBookingResponse'
          examples:
            success:
              value:
                message: Booking cancelled successfully
                booking:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  status: CANCELLED
                  cancelled_at: '2025-05-18T12:34:56.000Z'
        '400':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                invalid_token:
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid cancellation token
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                not_found:
                  value:
                    error:
                      code: BOOKING_NOT_FOUND
                      message: Booking not found
        '409':
          description: Booking is already cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                already_cancelled:
                  value:
                    error:
                      code: ALREADY_CANCELLED
                      message: Booking is already cancelled
  /api/calendar/ics:
    get:
      summary: Get calendar event as .ics file
      description: |
        Generate and download a .ics (iCalendar) file for a booking event. This file can be imported into any calendar application (Google Calendar, Outlook, Apple Calendar, etc.).

        The file contains:
        - Event title (client name)
        - Date and time
        - Studio information in description
        - Client contact information
      operationId: getCalendarIcs
      tags:
        - Calendar
      parameters:
        - name: booking_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Booking ID
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: .ics file generated successfully
          content:
            text/calendar:
              schema:
                type: string
                format: binary
              examples:
                success:
                  value: |
                    BEGIN:VCALENDAR
                    VERSION:2.0
                    PRODID:-//WOVSDH Nails//Booking//EN
                    BEGIN:VEVENT
                    UID:550e8400-e29b-41d4-a716-446655440000@wovsdh-nails.com
                    DTSTART:20250601T080000
                    DTEND:20250601T090000
                    SUMMARY:Анна Иванова
                    DESCRIPTION:Телефон: +972501234567\nEmail: anna@example.com
                    LOCATION:WOVSDH Nails Rishon
                    END:VEVENT
                    END:VCALENDAR
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                not_found:
                  value:
                    error:
                      code: BOOKING_NOT_FOUND
                      message: Booking not found
...
components:
  schemas:
...
    BookingCreatedDTO:
      type: object
      description: Minimal booking information returned after creation
      properties:
        id:
          type: string
          format: uuid
          description: Booking ID
        slot_id:
          type: string
          format: uuid
          description: Associated slot ID
        studio_id:
          type: string
          description: Studio ID
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - CANCELLED
          description: Booking status
        start_at:
          type: string
          format: date-time
          description: Event start time
        end_at:
          type: string
          format: date-time
          description: Event end time
        created_at:
          type: string
          format: date-time
          description: When booking was created
      required:
        - id
        - slot_id
        - studio_id
        - status
        - start_at
        - end_at
        - created_at
    CreateBookingRequest:
      type: object
      description: Request to create a new booking
      properties:
        slot_id:
          type: string
          format: uuid
          description: Slot to book
        studio_id:
          type: string
          description: Studio identifier
        client_first_name:
          type: string
          description: Client first name
        client_last_name:
          type: string
          description: Client last name
        client_phone:
          type: string
          description: Client phone number (including country code, e.g. +972...)
        client_email:
          type: string
          format: email
          description: Client email address
      required:
        - slot_id
        - studio_id
        - client_first_name
        - client_last_name
        - client_phone
        - client_email
    CancelBookingRequest:
      type: object
      description: Request to cancel a booking
      properties:
        token:
          type: string
          format: uuid
          description: Cancellation token from email
      required:
        - token
...
    CreateBookingResponse:
      type: object
      properties:
        booking:
          $ref: '#/components/schemas/BookingCreatedDTO'
    CancelBookingResponse:
      type: object
      properties:
        message:
          type: string
        booking:
          type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum:
                - CANCELLED
            cancelled_at:
              type: string
              format: date-time
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_PARAMS
                - SLOT_NOT_FOUND
                - SLOT_UNAVAILABLE
                - SLOT_HAS_ACTIVE_BOOKING
                - BOOKING_NOT_FOUND
                - ALREADY_CANCELLED
                - UNAUTHORIZED
                - VALIDATION_ERROR
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
...
  securitySchemes:
    AdminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin secret key (value of ADMIN_SECRET_KEY env var)

    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Internal secret key (value of ADMIN_SECRET_KEY env var)
...